requirements.txt
# Core ML libraries
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
peft>=0.7.0
accelerate>=0.25.0
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Data processing
numpy>=1.21.0
pandas>=1.3.0
psutil>=5.9.0

# Configuration
python-dotenv>=0.19.0

# Optional - for experiment tracking
# wandb>=0.15.0

# Optional - for better performance on some systems
# bitsandbytes>=0.41.0

# Development dependencies (optional)
# jupyter>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0 